+++ src/zfit/models/cache.py
@@ -10,13 +10,14 @@ import pydantic.v1 as pydantic
 import tensorflow as tf
 
 import zfit.z.numpy as znp
-from zfit.util.exception import AnalyticGradientNotAvailable
 
+from .. import z
 from ..core.serialmixin import SerializableMixin
 from ..core.space import supports
 from ..serialization import Serializer  # noqa: F401
 from ..settings import ztypes
 from ..util import ztyping
+from ..util.exception import AnalyticGradientNotAvailable
 from .basefunctor import FunctorPDFRepr
 from .functor import BaseFunctor
 
@@ -29,6 +30,7 @@ def get_value(cache: tf.Variable, flag: tf.Variable, func: Callable):
     def actual_func():
         def autoset_func():
             val = func()
+            # tf.print(val)
             return cache.assign(val, read_value=True)
 
         def use_cache():
@@ -53,6 +55,7 @@ class CachedPDF(BaseFunctor, SerializableMixin):
         self,
         pdf: ztyping.PDFInputType,
         *,
+        epsilon: float | None = None,
         extended: ztyping.ExtendedInputType = None,
         norm: ztyping.NormInputType = None,
         cache_tol=None,
@@ -63,10 +66,7 @@ class CachedPDF(BaseFunctor, SerializableMixin):
 
         .. note::
 
-           Analytic gradients are not available for the cached PDF. Use the numerical gradient instead,
-           either by using a minimizers internal calculator (e.g. :py:class:`~zfit.minimize.Minuit(..., gradient=True)`) or by
-           setting the autograd mode to False (e.g. :py:func:`~zfit.run.set_autograd_mode(False)`).
-           An error will be raised if the analytic gradient is requested.
+           Analytic gradients are not supported for cached PDFs. Use numerical gradients instead.
 
         The method stores the last calculated value of a function for a specific dataset and
         returns it when the input data and the parameters are the same. This can be useful when
@@ -92,11 +92,13 @@ class CachedPDF(BaseFunctor, SerializableMixin):
                Has no programmatical functional purpose as identification. |@docend:pdf.init.label|
         """
         obs = pdf.space
+        if cache_tol is not None:
+            epsilon = cache_tol
         hs3_init = {
             "obs": obs,
             "extended": extended,
             "norm": norm,
-            "cache_tol": cache_tol,
+            "epsilon": epsilon,
             "name": name,
         }
         name = name or pdf.name
@@ -104,13 +106,13 @@ class CachedPDF(BaseFunctor, SerializableMixin):
         params = list(pdf.get_params(floating=None, is_yield=None, extract_independent=True))
 
         param_cache = tf.Variable(
-            znp.zeros(shape=tf.shape(tf.stack(params)), dtype=ztypes.float),
+            znp.zeros(shape=tf.shape(znp.stack(params)), dtype=ztypes.float),
             trainable=False,
             validate_shape=False,
             dtype=tf.float64,
         )
         param_cache_int = tf.Variable(
-            znp.zeros(shape=tf.shape(tf.stack(params)), dtype=ztypes.float),
+            znp.zeros(shape=tf.shape(znp.stack(params)), dtype=ztypes.float),
             trainable=False,
             validate_shape=False,
             dtype=tf.float64,
@@ -126,10 +128,11 @@ class CachedPDF(BaseFunctor, SerializableMixin):
         self._integral_cache = None
         self._integral_cache_valid = tf.Variable(initial_value=False, trainable=False)
 
-        self._cache_tolerance = 1e-8 if cache_tol is None else cache_tol
+        self._cache_tolerance = 1e-8 if epsilon is None else epsilon
         self.hs3.original_init.update(hs3_init)
 
     @supports(norm="space")
+    @z.function(autograph=True)
     def _pdf(self, x, norm):
         x = x.value()
         xlen = tf.shape(x)[0]
@@ -141,6 +144,7 @@ class CachedPDF(BaseFunctor, SerializableMixin):
                 ),  # negative ones, to make sure these are unrealistic values
                 trainable=False,
                 validate_shape=False,
+                shape=tf.TensorShape([None]),
                 dtype=ztypes.float,
             )
 
@@ -149,15 +153,26 @@ class CachedPDF(BaseFunctor, SerializableMixin):
                 x + 19.0,  # to make sure it's not the same
                 trainable=False,
                 validate_shape=False,
+                shape=tf.TensorShape([None, x.shape[1]]),
                 dtype=ztypes.float,
             )
         cachedxlen = tf.shape(self._cached_x)[0]
         minlen = znp.min([xlen, cachedxlen])
-        x_same = tf.math.reduce_all(znp.abs(x[:minlen] - self._cached_x[:minlen]) < self._cache_tolerance)
+        xtrunc = x[:minlen]
+        xcachedtrunc = self._cached_x[:minlen]
+        # for debugging purposes, this fails a lot...
+        # tf.print(tf.shape(xtrunc))
+        # tf.print(tf.shape(xcachedtrunc))
+        # tf.print(xlen, cachedxlen, minlen)
+        with tf.control_dependencies([xtrunc, xcachedtrunc]):  # required! Otherwise would use previous shape (bug?)
+            xtrunc_diff = xtrunc - xcachedtrunc
+        xtruncdiff_abs = znp.abs(xtrunc_diff)
+        xtrunc_lt = xtruncdiff_abs < self._cache_tolerance
+        x_same = tf.math.reduce_all(xtrunc_lt)
         x_same = znp.logical_and(x_same, xlen == cachedxlen)
         pdf_params = list(self.pdfs[0].get_params())
         if hasparams := (paramlen := len(pdf_params)) > 0:
-            stacked_pdf_params = tf.stack(pdf_params)
+            stacked_pdf_params = znp.stack(pdf_params)
             cachedparamlen = tf.shape(self._cached_pdf_params)[0]
             minparamlen = znp.min(
                 [
@@ -165,10 +180,10 @@ class CachedPDF(BaseFunctor, SerializableMixin):
                     cachedparamlen,
                 ]
             )
-            params_same = tf.math.reduce_all(
-                znp.abs(stacked_pdf_params[:minparamlen] - self._cached_pdf_params[:minparamlen])
-                < self._cache_tolerance
-            )
+            stackedpdf_trunc = stacked_pdf_params[:minparamlen]
+            cachedpdf_trunc = self._cached_pdf_params[:minparamlen]
+            with tf.control_dependencies([stackedpdf_trunc, cachedpdf_trunc]):
+                params_same = tf.math.reduce_all(znp.abs(stackedpdf_trunc - cachedpdf_trunc) < self._cache_tolerance)
 
             same_args = tf.math.logical_and(params_same, x_same)
             same_args = znp.logical_and(same_args, paramlen == cachedparamlen)
@@ -182,8 +197,10 @@ class CachedPDF(BaseFunctor, SerializableMixin):
             self._cached_x.assign(x, read_value=False)
             return self.pdfs[0].pdf(x, norm)
 
+        # tf.print(self._pdf_cache)
         with tf.control_dependencies([assign1]):
-            return get_value(self._pdf_cache, self._pdf_cache_valid, value_update_func)
+            returnval = get_value(self._pdf_cache, self._pdf_cache_valid, value_update_func)
+        return returnval
 
     @supports(norm="space")
     def _integrate(self, limits, norm, options=None):
@@ -210,12 +227,12 @@ class CachedPDF(BaseFunctor, SerializableMixin):
 
         params = list(self.pdfs[0].get_params(floating=None))
         if hasparams := len(params) > 0:
-            stacked_pdf_params = tf.stack(params)
-            params_same = tf.math.reduce_all(
+            stacked_pdf_params = znp.stack(params)
+            params_same = znp.all(
                 znp.abs(stacked_pdf_params - self._cached_pdf_params_for_integration) < self._cache_tolerance
             )
 
-            same_args = tf.math.logical_and(params_same, limits_same)
+            same_args = znp.logical_and(params_same, limits_same)
         else:
             same_args = limits_same
         assign1 = self._integral_cache_valid.assign(same_args, read_value=False)